Allow "ksv" to use pre-compiled Ruby-files instead of compiling itself.
If one runs into problems with the compiled KSY and things break because
bugs/missing implementation in the compiler most likely, sometimes it's
easier to change once generated Ruby to test which changes lead to
successful Ruby and afterwards change the compiler accordingly. With the
former behaviour of "ksv" compiling always using temporary directories
this was pretty much impossible to achieve and one needed to change the
compiler always to generate new code. But for people new to Scala and
Ruby this is a pretty difficult roundtrip.

So "ksv" has been enhanced so that some pre-compiled Ruby file can be
given instead of some KSY and "ksv" doesn't compile in that case
anymore. Instead, by convention all Ruby files in the parent dir of the
given file are required and and the given file name is used as main
class. Both approaches are pretty much what "ksv" implemented currently
anyway.