A redraw isn't needed on Windows as it seems to be done anyway.
If the console on Windows is resized and even if one only uses ARROW
keys, the whole screen is corrected to look good again even without an
additional complete redraw like is issued in Tree.on_resize. with that
additional redraw things feel far slower than they need to be, most
likely because more than necessary is redrawn on each key invocation.